Moved TaskCompletedCard -> modules/todo/components. Renamed TaskCompletedList -> TaskCompletedCard. Did the same to the TaskControls component
